+++ mybatis/mybatis-lpus/src/main/java/cn/boot/mybatisplus/dataobject/Role.java
@@ -12,6 +12,6 @@ public class Role {
 
     private String name;
 
-    private Byte isDelete;
+    private Integer isDelete;
 
 }
