+++ mybatis/mybatis-lpus/src/main/java/cn/boot/mybatisplus/dataobject/User.java
@@ -14,7 +14,7 @@ public class User {
 
     private String password;
 
-    private Byte isDelete;
+    private Integer isDelete;
 
 
 }
